[mod] data binding 적용, home 패키지 분리
