+++ app/build.gradle
@@ -48,6 +48,8 @@ dependencies {
     androidTestImplementation 'androidx.test.ext:junit:1.1.5'
     androidTestImplementation 'androidx.test.espresso:espresso-core:3.5.1'
 
+    implementation "androidx.fragment:fragment-ktx:1.5.5"
+
     // Timber
     implementation 'com.jakewharton.timber:timber:5.0.1'
