+++ app/src/main/java/com/android/shoppi/presentation/home/HomeBannerAdapter.kt
@@ -1,0 +1,45 @@
+package com.android.shoppi.presentation.home
+
+import android.graphics.Color
+import android.graphics.drawable.ColorDrawable
+import android.view.LayoutInflater
+import android.view.ViewGroup
+import androidx.recyclerview.widget.DiffUtil
+import androidx.recyclerview.widget.ListAdapter
+import androidx.recyclerview.widget.RecyclerView
+import com.android.shoppi.databinding.ItemHomeBannerBinding
+import com.android.shoppi.presentation.main.Banner
+
+class HomeBannerAdapter : ListAdapter<Banner, HomeBannerAdapter.HomeBannerViewHolder>(
+    BannerDiffCallback()
+) {
+    class HomeBannerViewHolder(private val binding: ItemHomeBannerBinding) :
+        RecyclerView.ViewHolder(binding.root) {
+
+        fun bind(banner: Banner) {
+            binding.banner = banner
+            binding.tvBannerBadge.background =
+                ColorDrawable(Color.parseColor(banner.badge.backgroundColor))
+        }
+    }
+
+    override fun onCreateViewHolder(parent: ViewGroup, viewType: Int): HomeBannerViewHolder {
+        val inflater = LayoutInflater.from(parent.context)
+        return HomeBannerViewHolder(ItemHomeBannerBinding.inflate(inflater, parent, false))
+    }
+
+    override fun onBindViewHolder(holder: HomeBannerViewHolder, position: Int) {
+        holder.bind(getItem(position))
+    }
+}
+
+class BannerDiffCallback : DiffUtil.ItemCallback<Banner>() {
+    override fun areItemsTheSame(oldItem: Banner, newItem: Banner): Boolean {
+        return oldItem.productDetail.productId == newItem.productDetail.productId
+    }
+
+    override fun areContentsTheSame(oldItem: Banner, newItem: Banner): Boolean {
+        return oldItem == newItem
+    }
+
+}
