+++ app/src/main/res/navigation/main_navigation.xml
@@ -6,7 +6,7 @@
 
     <fragment
         android:id="@+id/navigation_home"
-        android:name="com.android.shoppi.presentation.main.HomeFragment"
+        android:name="com.android.shoppi.presentation.home.HomeFragment"
         android:label="HomeFragment" >
         <action
             android:id="@+id/action_home_to_product_detail"
