+++ app/src/main/res/values/strings.xml
@@ -7,4 +7,7 @@
     <string name="label_home_icon">홈</string>
     <string name="label_cart_icon">장바구니</string>
     <string name="home_info">Ivy님 이런 상품\n어때요?</string>
+
+    <string name="unit_discount_rate">%d%%</string>
+    <string name="unit_discount_currency">%s원</string>
 </resources>
